FEAT: Use a proxy for forwarding of Stream
